piGAN_lib/requirements.txt
numpy==1.19.5
pickleshare==0.7.5
Pillow==8.2.0
python-dateutil==2.8.1
pytorch-fid==0.2.0
pytorch-lightning==1.3.4
scikit-image==0.17.2
scikit-video==1.1.11
scipy==1.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
torch==1.8.1
torch-ema @ git+https://github.com/fadel/pytorch_ema@3950a7b5c4b88f46fd14f620277bad21898597a9
torch-fidelity==0.2.0
torchmetrics==0.2.0
torchvision==0.9.1
tqdm==4.60.0
traitlets==4.3.3